Let user reopen kvalitetsvurdering
