+++ frontend/src/components/kvalitetsregistrering/footer/back-link.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { NavLink } from 'react-router-dom';
+
+export const BackLink = () => (
+  <NavLink to="/kvalitetsregistreringer" className="knapp knapp--mini footer-button">
+    Tilbake
+  </NavLink>
+);
